parser/requirements.txt
beautifulsoup4
requests
flask
APScheduler
python-telegram-bot
python-dotenv